Add files via upload
In this repo, the fundamentals of the NumPy array in Python are examined and discussed in detail. 
Sufficient information for manipulating data using NumPy is presented, and the principles and requirements for broadcasting NumPy arrays are explained.
Also, importing images from local machines or online repositories to Python is shown.
Finally, converting imported images to NumPy arrays and getting the arrays ready for machine learning applications is discussed with clean code.
Enjoy!